minifix

+++ tasks/mpi/tsatsyn_a_increasing_contrast_by_histogram/src/ops_mpi.cpp
@@ -38,13 +38,13 @@ bool tsatsyn_a_increasing_contrast_by_histogram_mpi::TestMPITaskParallel::run() 
 
     for (int proc = 1; proc < world.size(); proc++) {
       local_data.clear();
-      for (int i = proc; i < input_data.size(); i += world.size()) {
+      for (int i = proc; i < static_cast<int>(input_data.size()); i += world.size()) {
         local_data.emplace_back(input_data[i]);
       }
       world.send(proc, 0, local_data);
     }
     local_data.clear();
-    for (int i = 0; i < input_data.size(); i += world.size()) {
+    for (int i = 0; i < static_cast<int>(input_data.size()); i += world.size()) {
       local_data.emplace_back(input_data[i]);
     }
   } else {
@@ -53,14 +53,14 @@ bool tsatsyn_a_increasing_contrast_by_histogram_mpi::TestMPITaskParallel::run() 
   boost::mpi::broadcast(world, height, 0);
   boost::mpi::broadcast(world, width, 0);
   std::vector<double> numbers(256, 0);
-  for (int i = 0; i < local_data.size(); i++) {
+  for (int i = 0; i < static_cast<int>(local_data.size()); i++) {
     numbers[local_data[i]]++;
   }
   if (world.rank() == 0) {
     std::vector<double> received_numbers(256, 0);
     for (int proc = 1; proc < world.size(); proc++) {
       world.recv(proc, 0, received_numbers);
-      for (int i = 0; i < numbers.size(); i++) {
+      for (int i = 0; i < static_cast<int>(numbers.size()); i++) {
         numbers[i] += received_numbers[i];
       }
     }
@@ -68,7 +68,7 @@ bool tsatsyn_a_increasing_contrast_by_histogram_mpi::TestMPITaskParallel::run() 
     world.send(0, 0, numbers);
   }
   if (world.rank() == 0) {
-    for (int i = 0; i < numbers.size(); i++) {
+    for (int i = 0; i < static_cast<int>(numbers.size()); i++) {
       numbers[i] /= width;
       std::cout << numbers[i] << " ";
     }
@@ -78,7 +78,7 @@ bool tsatsyn_a_increasing_contrast_by_histogram_mpi::TestMPITaskParallel::run() 
     std::cout << std::endl;
 
     std::cout << numbers[0] << " ";
-    for (int i = 1; i < numbers.size(); i++) {
+    for (int i = 1; i < static_cast<int>(numbers.size()); i++) {
       numbers[i] = (numbers[i] + numbers[i - 1]);
       std::cout << numbers[i] << " ";
     }
@@ -86,7 +86,7 @@ bool tsatsyn_a_increasing_contrast_by_histogram_mpi::TestMPITaskParallel::run() 
     std::cout << std::endl;
     std::cout << std::endl;
     std::cout << std::endl;
-    for (int i = 0; i < numbers.size(); i++) {
+    for (int i = 0; i < static_cast<int>(numbers.size()); i++) {
       numbers[i] *= 255;
       std::cout << numbers[i] << " ";
     }
@@ -94,7 +94,7 @@ bool tsatsyn_a_increasing_contrast_by_histogram_mpi::TestMPITaskParallel::run() 
     std::cout << std::endl;
     std::cout << std::endl;
     std::cout << std::endl;
-    for (int i = 0; i < numbers.size(); i++) {
+    for (int i = 0; i < static_cast<int>(numbers.size()); i++) {
       numbers[i] = round(numbers[i]);
       std::cout << numbers[i] << " ";
     }
